redesigned file layout and created classes via structures
