+++ src/ffmpeg/go.mod
@@ -1,0 +1,3 @@
+module github.com/gordon-cs/SIL-Video/Compiler/ffmpeg_pkg
+
+go 1.17
