+++ src/helper/go.mod
@@ -1,3 +1,3 @@
-module helper
+module github.com/gordon-cs/SIL-Video/Compiler/helper
 
 go 1.17
