+++ src/xml/go.mod
@@ -1,0 +1,3 @@
+module github.com/gordon-cs/SIL-Video/Compiler/xml
+
+go 1.17
